# src/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "VNPRI",
    "termsOfService": "",
    "contact": {
      "email": ""
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/notifications/": {
      "get": {
        "tags": ["Notifications"],
        "summary": "Get User Notifications",
        "description": "Get User Notifications",
        "parameters": [
          {
            "$ref": "#/components/parameters/AcceptLanguage"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/successResponseObject"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/badRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/forbiddenResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/noDataResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/serverErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "userAccessToken": []
          }
        ],
        "x-codegen-request-body-name": "body"
      }
    },
    "/notifications/read-all": {
      "post": {
        "tags": ["Notifications"],
        "summary": "Mark All Notifications As Read",
        "description": "Mark All Notifications As Read By User",
        "parameters": [
          {
            "$ref": "#/components/parameters/AcceptLanguage"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/successResponseObject"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/badRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/forbiddenResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/noDataResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/serverErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "userAccessToken": []
          }
        ],
        "x-codegen-request-body-name": "body"
      }
    },
    "/notifications/read-one/{id}": {
      "post": {
        "tags": ["Notifications"],
        "summary": "Mark One Notification As Read",
        "description": "Mark One Notification As Read By User",
        "parameters": [
          {
            "$ref": "#/components/parameters/AcceptLanguage"
          },
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/successResponseObject"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/badRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/forbiddenResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/noDataResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/serverErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "userAccessToken": []
          }
        ],
        "x-codegen-request-body-name": "body"
      }
    },
    "/notifications/save-device": {
      "post": {
        "tags": ["Notifications"],
        "summary": "Save Device Token",
        "description": "Save Device Token By User",
        "parameters": [
          {
            "$ref": "#/components/parameters/AcceptLanguage"
          }
        ],
        "requestBody": {
          "description": "An object defining our schema for this request",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/successResponseObject"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/badRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/forbiddenResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/noDataResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/serverErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "userAccessToken": []
          }
        ],
        "x-codegen-request-body-name": "body"
      }
    },
    "/notifications/settings": {
      "get": {
        "tags": ["Notifications"],
        "summary": "Get Notification Settings",
        "description": "Get Notification Settings By User",
        "parameters": [
          {
            "$ref": "#/components/parameters/AcceptLanguage"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/successResponseObject"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/badRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/forbiddenResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/noDataResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/serverErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "userAccessToken": []
          }
        ],
        "x-codegen-request-body-name": "body"
      },
      "post": {
        "tags": ["Notifications"],
        "summary": "Update Notification Settings",
        "description": "Update Notification Settings By User",
        "parameters": [
          {
            "$ref": "#/components/parameters/AcceptLanguage"
          }
        ],
        "requestBody": {
          "description": "An object defining our schema for this request",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "post_comment": {
                    "type": "boolean"
                  },
                  "comment_reply": {
                    "type": "boolean"
                  },
                  "comment_like": {
                    "type": "boolean"
                  },
                  "poll_vote": {
                    "type": "boolean"
                  },
                  "follow_post_activity": {
                    "type": "boolean"
                  },
                  "coin_transfer": {
                    "type": "boolean"
                  },
                  "direct_message": {
                    "type": "boolean"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/successResponseObject"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/badRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/forbiddenResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/noDataResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/serverErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "userAccessToken": []
          }
        ],
        "x-codegen-request-body-name": "body"
      }
    },
...
  },
  "components": {
    "schemas": {
      "serverErrorResponse": {
        "title": "Server Error Response",
        "type": "object",
        "properties": {
          "status": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "err": {
            "type": "string"
          }
        },
        "example": {
          "status": 500,
          "message": "Something Went Wrong",
          "err": "Error"
        }
      },
      "badRequestResponse": {
        "title": "Bad Request Response",
        "type": "object",
        "properties": {
          "status": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "validationErrors": {
            "type": "string"
          }
        },
        "example": {
          "status": 400,
          "message": "Something Went Wrong",
          "validationErrors": "Error"
        }
      },
      "forbiddenResponse": {
        "title": "Unauthorized Response",
        "type": "object",
        "properties": {
          "status": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        },
        "example": {
          "status": 401,
          "message": "Unauthorized"
        }
      },
      "noDataResponseObject": {
        "title": "No Data Object Response",
        "type": "object",
        "properties": {
          "status": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "string"
          }
        },
        "example": {
          "status": 404,
          "message": "No Data Found",
          "data": ""
        }
      },
...
      "successResponseObject": {
        "title": "Success Response",
        "type": "object",
        "properties": {
          "status": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {}
          }
        },
        "example": {
          "status": 200,
          "message": "Successful",
          "data": {}
        }
      }
    },
    "parameters": {
      "AcceptLanguage": {
        "name": "accept-language",
        "in": "header",
        "description": "Incase we want to send current language along our request",
        "schema": {
          "type": "string",
          "default": "en"
        }
      },
...
      "Id": {
        "name": "id",
        "required": true,
        "in": "path",
        "schema": {
          "type": "number"
        }
      },
...
    },
    "securitySchemes": {
      "userAccessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "userRefreshToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}